refactor(user/activity): split handle_incoming into per-verb submodules
Break the 1535-line handle_incoming.rs into a parent file holding only
the dispatch + shared helpers, plus one submodule per ActivityPub verb
(create, follow, like, announce, delete, undo). Tests and JSON builders
move alongside the handlers they exercise; shared builders live in
test_helpers.rs.
